moved homescreen directory
