+++ src/javascripts/main.js
@@ -3,7 +3,7 @@ import apiKeys from './helpers/apiKeys.json';
 import auth from './components/auth/auth';
 import authData from './helpers/data/authData';
 import listeners from './helpers/listeners';
-import homescreen from './components/ride/homescreen/homescreen';
+import homescreen from './components/homescreen/homescreen';
 
 import '../styles/main.scss';
 import 'bootstrap';
